+++ src/Model/Table/AppTable.php
@@ -14,6 +14,7 @@ namespace MeCms\Model\Table;
 
 use ArrayObject;
 use Cake\Cache\Cache;
+use Cake\Datasource\EntityInterface;
 use Cake\Event\Event;
 use Cake\I18n\FrozenTime;
 use Cake\I18n\Time;
@@ -36,31 +37,27 @@ class AppTable extends Table
     /**
      * Called after an entity has been deleted
      * @param \Cake\Event\Event $event Event object
-     * @param \Cake\ORM\Entity $entity Entity object
+     * @param \Cake\Datasource\EntityInterface $entity Entity object
      * @param \ArrayObject $options Options
      * @return void
-     * @uses getCacheName()
+     * @uses clearCache()
      */
-    public function afterDelete(Event $event, Entity $entity, ArrayObject $options)
+    public function afterDelete(Event $event, EntityInterface $entity, ArrayObject $options)
     {
-        if ($this->getCacheName()) {
-            Cache::clear(false, $this->getCacheName());
-        }
+        $this->clearCache();
     }
 
     /**
      * Called after an entity is saved
      * @param \Cake\Event\Event $event Event object
-     * @param \Cake\ORM\Entity $entity Entity object
+     * @param \Cake\Datasource\EntityInterface $entity Entity object
      * @param \ArrayObject $options Options
      * @return void
-     * @uses getCacheName()
+     * @uses clearCache()
      */
-    public function afterSave(Event $event, Entity $entity, ArrayObject $options)
+    public function afterSave(Event $event, EntityInterface $entity, ArrayObject $options)
     {
-        if ($this->getCacheName()) {
-            Cache::clear(false, $this->getCacheName());
-        }
+        $this->clearCache();
     }
 
     /**
@@ -88,20 +85,27 @@ class AppTable extends Table
     }
 
     /**
+     * Delete all keys from the cache
+     * @return bool `true` if the cache was successfully cleared, `false` otherwise
+     * @uses getCacheName()
+     */
+    public function clearCache()
+    {
+        return Cache::clear(false, $this->getCacheName());
+    }
+
+    /**
      * Deletes all records matching the provided conditions
      * @param mixed $conditions Conditions to be used, accepts anything
      *  `Query::where()` can take
      * @return int Returns the number of affected rows
+     * @uses clearCache()
      */
     public function deleteAll($conditions)
     {
-        $result = parent::deleteAll($conditions);
+        $this->clearCache();
 
-        if ($this->getCacheName()) {
-            Cache::clear(false, $this->getCacheName());
-        }
-
-        return $result;
+        return parent::deleteAll($conditions);
     }
 
     /**