+++ tests/TestCase/Model/Table/AppTableTest.php
@@ -77,11 +77,13 @@ class AppTableTest extends TableTestCase
      */
     public function testAfterDeleteAndAfterSave()
     {
-        foreach (['afterDelete', 'afterSave'] as $methodToCall) {
-            Cache::write('testKey', 'testValue', $this->Posts->getCacheName());
-            $this->Posts->$methodToCall(new Event(null), new Entity(), new ArrayObject());
-            $this->assertFalse(Cache::read('testKey', $this->Posts->getCacheName()));
-        }
+        $table = $this->getMockForModel('MeCms.Posts', ['clearCache']);
+        $table->expects($this->exactly(2))
+            ->method('clearCache');
+
+        list($event, $entity, $options) = [new Event(null), new Entity(), new ArrayObject()];
+        $table->afterDelete($event, $entity, $options);
+        $table->afterSave($event, $entity, $options);
     }
 
     /**
@@ -121,13 +123,26 @@ class AppTableTest extends TableTestCase
     }
 
     /**
+     * Test for `clearCache()` method
+     * @test
+     */
+    public function testClearCache()
+    {
+        Cache::write('testKey', 'testValue', $this->Posts->getCacheName());
+        $this->assertTrue($this->Posts->clearCache());
+        $this->assertFalse(Cache::read('testKey', $this->Posts->getCacheName()));
+    }
+
+    /**
      * Test for `deleteAll()` method
      * @test
      */
     public function testDeleteAll()
     {
+        $this->assertNotEmpty($this->Posts->find()->count());
         Cache::write('testKey', 'testValue', $this->Posts->getCacheName());
-        $this->Posts->deleteAll(['id IS NOT' => null]);
+        $this->assertGreaterThan(0, $this->Posts->deleteAll(['id IS NOT' => null]));
+        $this->assertEmpty($this->Posts->find()->count());
         $this->assertFalse(Cache::read('testKey', $this->Posts->getCacheName()));
     }